Print free objects

Print methods for free objects

Usage

# S3 method for class 'free'
print(x,...)
as.character_free(m,latex=getOption("latex"))

Arguments

x

Object of class free in the print method

m

A two-row matrix in function as.character_free()

latex

Boolean, with TRUE meaning to print latex-friendly output including curly braces, and default NULL option meaning to give a nicer-looking output that latex would typeset incorrectly

...

Further arguments, currently ignored

Author

Robin K. S. Hankin

Note

The print method does not change the internal representation of a free object, which is a list of integer matrices.

The default print method uses multiplicative notation (powers) which is inconsistent with the juxtaposition method “+”.

The print method has special dispensation for length-zero free objects but these are not handled entirely consistently.

The default print method uses lowercase letters a-z, but it is possible to override this using options("freegroup_symbols" = foo), where foo is a character vector. This is desirable if you have more than 26 symbols, because unallocated symbols appear as NA.

The package will allow the user to set options("freegroup_symbols") to unhelpful things like rep("a",20) without complaining (but don't actually do it, you crazy fool).

See also

Examples

## default symbols:

abc(26)
#> [1] a.b.c.d.e.f.g.h.i.j.k.l.m.n.o.p.q.r.s.t.u.v.w.x.y.z
rfree(1,10)
#> [1] g^5.i^-4.d^9.i^3.a^5.h^9.j^-9.g^2.f^5.d^-4

# if we need more than 26:
options(freegroup_symbols=state.name)
rfree(10,4)
#>  [1] Arizona^-2.Alaska^-3                      
#>  [2] Alaska^-4.Arizona^-2.Arkansas^-4          
#>  [3] Alabama^2                                 
#>  [4] Arkansas^2.Alabama^2.Arkansas^-2.Alabama^4
#>  [5] Arkansas^-1.Alabama.Arkansas.Alabama^-2   
#>  [6] Alaska^7.Arkansas^-4.Arizona^2            
#>  [7] Alabama^-2.Arkansas^-3.Alaska             
#>  [8] Alabama^-4.Arkansas^-2.Arizona^-4         
#>  [9] Arizona^-1.Alaska^3                       
#> [10] Alabama^4.Alaska^-2.Arizona^4             

# or even:
jj <- letters[1:10]
options(freegroup_symbols=apply(expand.grid(jj,jj),1,paste,collapse=""))
rfree(10,10,100,4)
#>  [1] fe.fb^4.fi^3.hc^4.cc.ci^4.aj.he^2.ea^2.fc^4  
#>  [2] ac^2.ei^3.bh^4.df.cb.be^2.ei.bb.bf^2.ej^3    
#>  [3] eb^2.ig^4.je^4.hj^2.be^3.fb^2.eh.id^3.hd^2.ji
#>  [4] je.bi.ja^2.ce^3.ca^3.ac^2.gc^4.ib^3.di^2.hj^2
#>  [5] ih^2.jf^4.ce^2.ah.ej^2.ci^3.hi^2.fd.he^2.hg^4
#>  [6] ch^3.aj^4.eh^2.id.hb^4.gc.fh^2.ad^4.gg^3.cg^2
#>  [7] hb^4.ce^4.cf^3.ed^3.dc^3.dg^4.bg.ig^3.ai.bh^4
#>  [8] cg^3.hb^4.bg^3.ea.jg.eg^4.ca^4.ci^4.hd^4.ib  
#>  [9] jd.dj.ga^3.ee.jd^2.gb^3.ji.ib^3.ff^4.ej^3    
#> [10] hj^4.jj^3.eg^3.bd^3.cf^3.ig.hd.ed.dj^3.af^3  

options(freegroup_symbols=NULL)  #  NULL is interpreted as letters a-z
rfree(10,4)            #  back to normal
#>  [1] a^-1.d^2.a^3      c^-2.b^2.a^4      c^-2.d^-2.b^3     d^-2.c^-4.b^3    
#>  [5] d^-3              d^-1.b^4.c^-3     d^4.c^-3.b^-3.d^4 d^-2             
#>  [9] b^4.c.b^-4        a^-3.d^3.a^-4